In Guide folder: changed shared folder name to components, and moved PerformanceBoxRow tsx and scss into a folder with an index.tsx
