+++ src/analysis/retail/druid/feral/modules/spells/RakeUptimeAndSnapshots.tsx
@@ -15,7 +15,7 @@ import Snapshots, {
 import { TALENTS_DRUID } from 'common/TALENTS';
 import { proccedBloodtalons } from 'analysis/retail/druid/feral/normalizers/BloodtalonsLinkNormalizer';
 import { SubSection } from 'interface/guide';
-import { PerformanceBoxRow } from 'interface/guide/shared/PerformanceBoxRow';
+import { PerformanceBoxRow } from 'interface/guide/components/PerformanceBoxRow';
 import { QualitativePerformance } from 'parser/ui/QualitativePerformance';
 
 /** Tracking code for everything Rake related */
