+++ src/analysis/retail/druid/feral/modules/spells/RipUptimeAndSnapshots.tsx
@@ -28,7 +28,7 @@ import RESOURCE_TYPES from 'game/RESOURCE_TYPES';
 import { QualitativePerformance } from 'parser/ui/QualitativePerformance';
 import { SubSection } from 'interface/guide';
 import { SpellLink } from 'interface';
-import { PerformanceBoxRow } from 'interface/guide/shared/PerformanceBoxRow';
+import { PerformanceBoxRow } from 'interface/guide/components/PerformanceBoxRow';
 
 class RipUptimeAndSnapshots extends Snapshots {
   static dependencies = {
