+++ src/analysis/retail/monk/brewmaster/modules/problems/PurifyingBrew/index.tsx
@@ -5,8 +5,8 @@ import SPELLS from 'common/SPELLS';
 import { WclTable, WCLThreatBand, WCLThreatTableResponse } from 'common/WCL_TYPES';
 import { TooltipElement } from 'interface';
 import { GuideProps, SubSection } from 'interface/guide';
-import type { Problem, ProblemRendererProps } from 'interface/guide/shared/ProblemList';
-import ProblemList from 'interface/guide/shared/ProblemList';
+import type { Problem, ProblemRendererProps } from 'interface/guide/components/ProblemList';
+import ProblemList from 'interface/guide/components/ProblemList';
 import SpellLink from 'interface/SpellLink';
 import { AddStaggerEvent, EventType, RemoveStaggerEvent } from 'parser/core/Events';
 import { Info } from 'parser/core/metric';
@@ -22,7 +22,7 @@ import { potentialStaggerEvents } from './solver';
 
 import './PurifyingBrew.scss';
 import talents from 'common/TALENTS/monk';
-import PassFailBar from 'interface/guide/shared/PassFailBar';
+import PassFailBar from 'interface/guide/components/PassFailBar';
 
 export { default } from './analyzer';
