+++ src/interface/guide/components/Apl/violations/index.tsx
@@ -5,7 +5,7 @@ import ProblemList, {
   NoProblem,
   Problem,
   ProblemRendererProps,
-} from 'interface/guide/shared/ProblemList';
+} from 'interface/guide/components/ProblemList';
 import { Apl, CheckResult, Violation } from 'parser/shared/metrics/apl';
 import React, { useMemo } from 'react';
 import { useContext } from 'react';
@@ -17,7 +17,7 @@ import {
   ViolationExplainer,
 } from './claims';
 import deduplicate, { DEDUP_WINDOW } from './deduplication';
-import PassFailBar from 'interface/guide/shared/PassFailBar';
+import PassFailBar from 'interface/guide/components/PassFailBar';
 
 const EmbedContainer = styled.div`
   background: #222;
