+++ src/interface/guide/components/GradiatedPerformanceBar.tsx
