+++ src/interface/guide/components/PassFailBar.tsx
