+++ src/interface/guide/components/PerformanceBoxRow/PerformanceBoxRow.scss
@@ -1,4 +1,4 @@
-@import '../Guide';
+@import 'interface/guide/Guide';
 
 $shadow: 0 0 0px 1px hsla(45, 100%, 60%, 0.7);
